+++ src/Notemac/Commons/Enums.ts
@@ -16,4 +16,6 @@ export type RenderWhitespace = 'none' | 'boundary' | 'selection' | 'trailing' | 
 
 export type LineEnding = 'LF' | 'CRLF' | 'CR';
 
-export type ThemeName = 'mac-glass' | 'dark' | 'light' | 'monokai' | 'solarized-dark' | 'solarized-light' | 'dracula';
+export type ThemeName = 'mac-glass' | 'dark' | 'light' | 'monokai' | 'solarized-dark' | 'solarized-light' | 'dracula' | 'custom';
+
+export type BuiltInThemeName = 'mac-glass' | 'dark' | 'light' | 'monokai' | 'solarized-dark' | 'solarized-light' | 'dracula';